+++ src/app/[productId]/page.tsx
@@ -67,7 +67,7 @@ export default function ItemDisplay({
           <img
             src={Item?.photo}
             alt={Item?.name}
-            style={{ width: '600px', height: '600px' }}
+            style={{ width: '350px', height: '350px' }}
           />
         </ImageContainer>
         <TextContainer>
